+++ CHANGELOG.md
@@ -2,6 +2,14 @@
 
 All notable changes to `Weasley` will be documented in this file.
 
+## [0.6.0](https://github.com/rougin/weasley/compare/v0.5.0...v0.6.0) - Unreleased
+
+### Added
+- `Session` library
+
+### Changed
+- `Generator` directory to `Console` directory
+
 ## [0.5.0](https://github.com/rougin/weasley/compare/v0.4.4...v0.5.0) - 2017-10-06
 
 ### Changed